Feature: alternative approach to perturbative trajectories
Previously, the noise applied to branches in the perturbative expansion strategy traces the trajectory by ignoring branches having more errors than the maximum order. This led to the total probabilty being <1 when the order was less than the maximum order. This commit implements an alternative method of approximation where a branch is terminated when the number of errors applied in that branch reaches the maximum order limit. This should lead to the total probability being 1.

The previous version of the trajectory allows one to estimate the amount of approximation, that is related to the number of branches ignored, by calculating how far the total probability is, from 1. This option is missing in the new approach.