remove PauliFrameCircuitHandler
There is probably a place for a similar function, but first we need to
finish the basic building blocks, so that we have a clean API.
Similarly important is to not just introduce a new functions when there
are already parts of the API that implement similar features.

The doc strings also need to be updated, but I am leaving this for a
separate commit.