defining the main API
I was having a lot of trouble deciding on function names and API

There are already two functions for circuit simulation in this library, the mctrajectories and the petrajectories. There is some overlap in style between the already available mctrajectories and what you have implemented, thus my suggestion to create a new method for that function. But on second though, I now think it would be best to simply have 3 separate functions for the 3 separate styles of simulation. However, I am trying to have names that follow the same style, to be more consistent.

More changes to the way this new function is being used to follow.